tests/driver_at86rf2xx: rewrite without GNRC
