+++ tests/driver_at86rf2xx/Makefile.ci
@@ -3,9 +3,8 @@ BOARD_INSUFFICIENT_MEMORY := \
     arduino-leonardo \
     arduino-nano \
     arduino-uno \
-    atmega328p \
     atmega328p-xplained-mini \
-    nucleo-f031k6 \
+    atmega328p \
     nucleo-l011k4 \
     samd10-xmini \
     stm32f030f4-demo \
