+++ tests/driver_at86rf2xx/README.md
@@ -1,16 +1,16 @@
 # About
-This is a manual test application for the AT86RF2xx radio driver
+This is a manual test application for the AT86RF2xx radio driver.
 
 For running this test, you need to connect/configure the following pins of your
 radio device:
 - SPI MISO
 - SPI MOSI
 - SPI CLK
-- CS (ship select)
+- CS (chip select)
 - RESET
 - SLEEP
 - INT (external interrupt)
 
 # Usage
-For testing the radio driver you can use the netif and txtsnd shell commands
+For testing the radio driver you can use the ifconfig and txtsnd shell commands
 that are included in this application.
